feat: refactored code
